fix: Phase 5 compilation bugs — g_stats null ptr, lambda captures, enum casts, missing includes
Critical fixes:
- WebHandlers: g_stats pointer was never initialized (null pointer dereference on every action)
- WebHandlers: Added GameStats& parameter to registerHandlers() and forward declarations
- OTA: Added [&server] capture to all lambdas, added #include <Esp.h>
- WiFiManager: Added [&server] capture to all lambdas, added #include <Esp.h>
- MultiPet: Fixed invalid enum conversion (PetStage/PetType bitwise OR)
- Storage: Fixed invalid enum conversion (PetStage/PetType bitwise OR)
- Pet: Added #include <ArduinoJson.h>, removed duplicate getBatteryJson()
- Pet: Added missing NOTE_D5/NOTE_E5/NOTE_F5/NOTE_G5 definitions
- Stats: Removed duplicate STATS_FILE definition (already in config.h)
- Removed orphaned Statistics.cpp (duplicate of Stats.cpp)
- Added .gitignore for .pio/ build artifacts

Compilation: ✅ SUCCESS (RAM 16%, Flash 70.5%)

+++ src/MultiPet.cpp
@@ -33,13 +33,13 @@ void loadMultiPetState(MultiPetState &state) {
     p.isAlive     = obj["isAlive"] | true;
     p.state       = obj["state"].as<String>();
     if (p.state.length() == 0) p.state = "normal";
-    p.stage           = (PetStage)(int)obj["stage"] | BABY;
+    p.stage           = (PetStage)(int)(obj["stage"] | (int)BABY);
     p.isNight         = obj["isNight"] | false;
     p.virtualMinutes  = obj["virtualMinutes"] | 0UL;
     p.name            = obj["name"].as<String>();
     if (p.name.length() == 0) p.name = "Pet" + String(idx + 1);
     p.soundEnabled    = obj["soundEnabled"] | true;
-    p.type            = (PetType)(int)obj["type"] | BLOB;
+    p.type            = (PetType)(int)(obj["type"] | (int)BLOB);
     p.feedCount       = obj["feedCount"] | 0;
     p.playCount       = obj["playCount"] | 0;
     p.hasBeenNamed    = obj["hasBeenNamed"] | false;
